requirements.txt
flask==2.0.1
mysql-connector-python==8.0.26
sentence-transformers==2.2.2
faiss-cpu==1.7.2
flask-cors==3.0.10
python-dotenv==0.19.2
google-generativeai==0.1.0